+++ apps/ccoverview/vite.config.ts
@@ -1,11 +1,21 @@
 import * as path from 'path';
 import EnvironmentPlugin from 'vite-plugin-environment';
+import injectProcessEnv from 'vite-plugin-environment';
 export default {
   plugins: [EnvironmentPlugin({ NODE_ENV: 'development' })],
   resolve: {
     alias: {
       '@cc-components/ccoverviewapp': path.resolve('../../libs/ccoverviewapp/src'),
       '@cc-components/shared': path.resolve('../../libs/shared/src'),
+    },
+  },
+  build: {
+    rollupOptions: {
+      plugins: [
+        injectProcessEnv({
+          NODE_ENV: 'development',
+        }),
+      ],
     },
   },
 };