+++ apps/jcainstallation/vite.config.ts
@@ -1,7 +1,9 @@
 import * as path from 'path';
+import EnvironmentPlugin from 'vite-plugin-environment';
 import injectProcessEnv from 'rollup-plugin-inject-process-env';
 
 export default {
+  plugins: [EnvironmentPlugin({ NODE_ENV: 'development' })],
   resolve: {
     alias: {
       '@cc-components/jcainstallationapp': path.resolve(