+++ apps/loop/vite.config.ts
@@ -1,6 +1,6 @@
 import * as path from 'path';
 import EnvironmentPlugin from 'vite-plugin-environment';
-
+import injectProcessEnv from 'rollup-plugin-inject-process-env';
 import { Plugin } from 'vite-plugin-cdn-import';
 
 export default {
@@ -25,6 +25,15 @@ export default {
     alias: {
       '@cc-components/loopapp': path.resolve('../../libs/loopapp/src'),
       '@cc-components/shared': path.resolve('../../libs/shared/src'),
+    },
+  },
+  build: {
+    rollupOptions: {
+      plugins: [
+        injectProcessEnv({
+          NODE_ENV: 'development',
+        }),
+      ],
     },
   },
 };