+++ apps/mechanicalcompletion/vite.config.ts
@@ -1,5 +1,6 @@
 import * as path from 'path';
 import EnvironmentPlugin from 'vite-plugin-environment';
+import injectProcessEnv from 'rollup-plugin-inject-process-env';
 export default {
   plugins: [EnvironmentPlugin({ NODE_ENV: 'development' })],
   resolve: {
@@ -14,6 +15,15 @@ export default {
       '@cc-components/mechanicalcompletionsidesheet': path.resolve(
         '../../libs/mechanicalcompletionsidesheet/src'
       ),
+    },
+  },
+  build: {
+    rollupOptions: {
+      plugins: [
+        injectProcessEnv({
+          NODE_ENV: 'development',
+        }),
+      ],
     },
   },
 };