+++ apps/preservationanalytics/vite.config.ts
@@ -1,5 +1,6 @@
 import * as path from 'path';
 import EnvironmentPlugin from 'vite-plugin-environment';
+import injectProcessEnv from 'rollup-plugin-inject-process-env';
 export default {
   plugins: [EnvironmentPlugin({ NODE_ENV: 'development' })],
   resolve: {
@@ -8,6 +9,15 @@ export default {
         '../../libs/preservationanalyticsapp/src'
       ),
       '@cc-components/shared': path.resolve('../../libs/shared/src'),
+    },
+  },
+  build: {
+    rollupOptions: {
+      plugins: [
+        injectProcessEnv({
+          NODE_ENV: 'development',
+        }),
+      ],
     },
   },
 };