+++ apps/punch/vite.config.ts
@@ -1,4 +1,5 @@
 import * as path from 'path';
+import injectProcessEnv from 'rollup-plugin-inject-process-env';
 export default {
   resolve: {
     alias: {
@@ -6,6 +7,15 @@ export default {
       '@cc-components/shared': path.resolve('../../libs/shared/src'),
       '@cc-components/punchshared': path.resolve('../../libs/punchshared/src'),
       '@cc-components/punchsidesheet': path.resolve('../../libs/punchsidesheet/src'),
+    },
+  },
+  build: {
+    rollupOptions: {
+      plugins: [
+        injectProcessEnv({
+          NODE_ENV: 'development',
+        }),
+      ],
     },
   },
 };