+++ apps/swcr/vite.config.ts
@@ -1,5 +1,6 @@
 import * as path from 'path';
 import EnvironmentPlugin from 'vite-plugin-environment';
+import injectProcessEnv from 'rollup-plugin-inject-process-env';
 
 export default {
   plugins: [EnvironmentPlugin({ NODE_ENV: 'development' })],
@@ -9,6 +10,15 @@ export default {
       '@cc-components/shared': path.resolve('../../libs/shared/src'),
       '@cc-components/swcrshared': path.resolve('../../libs/swcrshared/src'),
       '@cc-components/swcrsidesheet': path.resolve('../../libs/swcrsidesheet/src'),
+    },
+  },
+  build: {
+    rollupOptions: {
+      plugins: [
+        injectProcessEnv({
+          NODE_ENV: 'development',
+        }),
+      ],
     },
   },
 };