+++ apps/activities/CHANGELOG.md
@@ -1,3 +1,6 @@
+## Version 0.0.3
+Inject process.env to bundle
+
 ## Version 0.0.2
 Bump fusion-framework packages