+++ apps/loop/CHANGELOG.md
@@ -1,3 +1,6 @@
+## Version 0.0.9
+Inject process.env to bundle
+
 ## Version 0.0.8
 Bump fusion-framework packages