+++ apps/workorder/CHANGELOG.md
@@ -1,3 +1,6 @@
+## Version 0.0.8
+Inject process.env to bundle
+
 ## Version 0.0.7
 Bump fusion-framework packages